feat: add edit and create screen of obituaries
