+++ app/obituaries/index.tsx
@@ -17,15 +17,16 @@ const width = Dimensions.get("window").width;
 const height = Dimensions.get("window").height;
 
 type RootStackParamList = {
-  'obituaries/createObituary': {
-    imageTemplateId: number,
-    imageUrl: string,
-    is_newObituary: boolean,
-    obituaryId: number,
-    jsonData: string,
-    is_mine: boolean,
-    selectedColor: string,
-  };
+  // 'obituaries/createObituary': {
+  //   imageTemplateId: number,
+  //   imageUrl: string,
+  //   is_newObituary: boolean,
+  //   obituaryId: number,
+  //   jsonData: string,
+  //   is_mine: boolean,
+  //   selectedColor: string,
+  //   changeDesign: boolean,
+  // };
   'obituaries/listMyObituaries': undefined;
   'obituaries/index': {
     is_newObituary: boolean,
@@ -35,11 +36,26 @@ type RootStackParamList = {
     is_mine: boolean,
     selectedColor: string,
   };
-  'obituaries/makeObituary': {
+  'obituaries/createObituary': {
+    imageTemplateId: number,
     imageUrl: string,
     is_newObituary: boolean,
     obituaryId: number,
-    is_mine: boolean | undefined,
+    is_mine: boolean,
+    jsonData: string | undefined,
+    changeDesign: boolean,
+    selectedColor: string | undefined,
+  };
+
+  'obituaries/editObituary': { 
+    imageUrl: string; 
+    imageTemplateId: number; 
+    is_newObituary: boolean;
+    obituaryId: number;
+    is_mine: boolean | undefined;
+    jsonData: string | undefined;
+    changeDesign: boolean;
+    selectedColor: string | undefined;
   };
 
 };
@@ -58,8 +74,8 @@ function ObituaryIndex() {
   const [modalVisible, setModalVisible] = useState(false);
   const [modalMessage, setModalMessage] = useState("");
 
-  
 
+  
   interface Obituary {
     id: number
     imageId: number;
@@ -97,7 +113,6 @@ function ObituaryIndex() {
   );
 
   const showConfirmationModal = (id: number, imageUrl: string) => {
-
     setSelectedObituary({ id, imageUrl });
     setModalMessage("¿Para quién es la esquela?");
     setModalVisible(true);
@@ -124,27 +139,33 @@ function ObituaryIndex() {
       }
     }
 
-    const { obituaryId, jsonData, selectedColor } = route.params ?? {};
-    navigation.navigate('obituaries/makeObituary', {
+    const { obituaryId } = route.params ?? {};
+    navigation.navigate('obituaries/createObituary', {
+      imageTemplateId: selectedObituary.id,
       imageUrl: selectedObituary.imageUrl,
       is_newObituary,
       obituaryId,
       is_mine: true,
+      jsonData: undefined,
+      changeDesign: false,
+      selectedColor: undefined,
     });
-
     setModalVisible(false);
-
   }
 
   const handleElseObituary = async () => {
     if (!selectedObituary) return;
 
-    const { obituaryId, jsonData, selectedColor } = route.params ?? {};
-    navigation.navigate('obituaries/makeObituary', {
+    const { obituaryId } = route.params ?? {};
+    navigation.navigate('obituaries/createObituary', {
+      imageTemplateId: selectedObituary.id,
       imageUrl: selectedObituary.imageUrl,
       is_newObituary,
       obituaryId,
       is_mine: false,
+      jsonData: undefined,
+      changeDesign: false,
+      selectedColor: undefined
     });
 
     setModalVisible(false);
@@ -190,13 +211,31 @@ function ObituaryIndex() {
               key={item.id}
               onPress={() => {
                 if (changeDesign) {
-                  const { obituaryId, jsonData, is_mine, selectedColor } = route.params ?? {};
-                  navigation.navigate('obituaries/makeObituary', {
-                    imageUrl: item.imageUrl,
-                    is_newObituary,
-                    obituaryId,
-                    is_mine,
-                  });
+                  console.log("changeDesign", changeDesign);
+                  const { is_newObituary, obituaryId, jsonData, is_mine, selectedColor } = route.params ?? {};
+                  if (!is_newObituary) {
+                    navigation.navigate('obituaries/editObituary', {
+                      imageTemplateId: item.imageId,
+                      imageUrl: item.imageUrl,
+                      is_newObituary,
+                      obituaryId,
+                      is_mine,
+                      jsonData,
+                      changeDesign: true,
+                      selectedColor,
+                    });
+                  } else { 
+                    navigation.navigate('obituaries/createObituary', {
+                      imageTemplateId: item.imageId,
+                      imageUrl: item.imageUrl,
+                      is_newObituary,
+                      obituaryId,
+                      is_mine,
+                      jsonData,
+                      changeDesign: true,
+                      selectedColor,
+                    });
+                  }
                 } else {
                   showConfirmationModal(item.id, item.imageUrl);
                 }
